Move GetUseLocation to paths package
